Consolidate function files into utils.py
